+++ app/src/main/java/com/stedi/quizapp/model/QuizDetails.kt
@@ -3,6 +3,6 @@ package com.stedi.quizapp.model
 import com.google.gson.annotations.SerializedName
 
 data class QuizDetails(
-   @SerializedName("id") val id: Int? = null,
+   @SerializedName("id") val id: Double? = null,
    @SerializedName("questions") val questions: List<Question>? = null
 )
